+++ babel.config.js
@@ -22,8 +22,10 @@ module.exports = function (api) {
 
   const plugins = [
     ["inline-json-import", {}],
+    ["@babel/plugin-transform-runtime"],
     ["@babel/plugin-proposal-private-methods", { loose: true }],
     ["@babel/plugin-proposal-class-properties", { loose: true }],
+    ["@babel/plugin-proposal-private-property-in-object", { loose: true }],
   ];
 
   return {