+++ packages/naksha-gmaps-draw/stories/naksha-gmaps-draw.stories.tsx
@@ -7,6 +7,7 @@ export const NakshaGmapsDrawStory = () => (
     <NakshaGmapsDraw
       isAutocomplete={true}
       isMultiple={true}
+      isImport={true}
       gmapApiAccessToken={process.env.STORYBOOK_GMAP_TOKEN}
       // Restricts autocomplete + customization searches to India
       gmapRegion="IN"