+++ packages/naksha-gmaps-view/src/utils/geojson.tsx
@@ -1,0 +1,18 @@
+import bbox from "@turf/bbox";
+
+/**
+ * Calculates bounds for all features
+ *
+ * @param {*} geojson
+ * @return {*}
+ */
+export const calculateBounds = (geojson) => {
+  const b = bbox(geojson);
+
+  return {
+    east: b[2],
+    north: b[3],
+    south: b[1],
+    west: b[0],
+  };
+};