+++ packages/naksha-upload/src/icons/constants.tsx
@@ -2,6 +2,19 @@ export const FILE_TYPES = {
   DBF: ".dbf",
   SHP: ".shp",
   SHX: ".shx",
+  PRJ: ".prj",
+  SBN: ".sbn",
+  SBX: ".sbx",
+  FBN: ".fbn",
+  FBX: ".fbx",
+  AIN: ".ain",
+  AIH: ".aih",
+  IXS: ".ixs",
+  MXS: ".mxs",
+  ATX: ".atx",
+  XML: ".xml",
+  CPG: ".cpg",
+  QIX: ".qix",
 };
 
 export const LAYER_TYPES = [